Split the gramps module up in submodules to allow for growth and easier understanding and maintenance of the codebase.
